requirements.txt
Django==1.10.5
django-redis==4.7.0
django-widget-tweaks==1.4.1
-e git+https://github.com/MUCI/usosapi-python@74482224f678a6ceb038563920e02c23e93cebdc#egg=usosapi
tabula-py
